remove another useless cast

+++ libraries/ESP8266WiFi/src/WiFiClient.cpp
@@ -197,7 +197,7 @@ bool WiFiClient::getSync() const
 
 int WiFiClient::availableForWrite ()
 {
-    return _client? (int)_client->availableForWrite(): 0;
+    return _client? _client->availableForWrite(): 0;
 }
 
 size_t WiFiClient::write(uint8_t b)
